requirements.txt
click==7.0
flake8==3.7.7
pycotap==1.1.0
pylint==2.3.1
pytest==4.3.0
safety==1.8.5
gitpython==2.1.11
